% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctest.R
\name{mcr.test}
\alias{mcr.test}
\title{Multiple Correlation Test of Significance}
\usage{
mcr.test(x, ld, rd)
}
\arguments{
\item{x}{Data Matrix or Variance Covariance or Correlation matrix}

\item{ld}{Label of dependent Variable}

\item{rd}{Vector of labels of independent variables}
}
\value{
a htest class object
}
\description{
Tests the significance of mutliple correlation coefficient
}
\examples{
## Example
library(MASS)
mu<-c(10,12,13,14)
sig<-matrix(0,4,4)
diag(sig)<-c(2,1,1,2)
da<-mvrnorm(25,mu,sig)
mcr.test(da,1,c(2:4))
}
\author{
Abirami S
}
